backend/requirements.txt
fastapi==0.121.2
uvicorn[standard]==0.38.0
playwright==1.56.0
python-dotenv==1.0.1
pydantic==2.12.4
